requirements.txt
python-dotenv
langchain
pinecone>=3.0.0
langchain-groq
sentence-transformers
fastapi
uvicorn
python-multipart
python-dotenv
langchain-groq
sentence-transformers
streamlit
requests